requirements.txt
chromadb==1.5.8
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-core==1.3.2
langchain-huggingface==1.2.2
langchain-openrouter==0.2.1
langchain-text-splitters==1.1.2
langgraph==1.1.10
pytest==9.0.3
python-dotenv==1.2.2
requests==2.33.1
sentence-transformers==5.4.1
